pyproject.toml
[tool.poetry]
name = "technical-task-by-teamstroyit"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "technical_task_by_teamstroyit"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.27.1"
fastapi = "^0.110.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"
psycopg2-binary = "^2.9.9"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
